+++ integration-tests/tests/src/setup-globals.ts
@@ -62,3 +62,8 @@ describe("Test Harness", function (this: Mocha.Suite) {
   Suite.prototype.longTimeout = longTimeout;
   Context.prototype.longTimeout = longTimeout;
 });
+
+import { Realm } from "realm";
+
+//Disable the logger to avoid console flooding
+Realm.setLogLevel("off");