+++ packages/realm/src/index.ts
@@ -88,6 +88,7 @@ export {
   InvalidateEvent,
   List,
   LocalAppConfiguration,
+  LoggerCallback,
   LogLevel,
   mapTo,
   MapToDecorator,